+++ src/Sample/Quantization.java
@@ -3,7 +3,7 @@ package Sample;
 public class Quantization{
 
 	// Luminance Quantization Tables
-	private static double[][] lqt_recommended_luminance = {
+	private static float[][] lqt_recommended_luminance = {
 			{16, 11, 10, 16, 24,  40,  51,  61},
 			{12, 12, 14, 19, 26,  58,  60,  55},
 			{14, 13, 16, 24, 40,  57,  69,  56},
@@ -13,7 +13,7 @@ public class Quantization{
 			{49, 64, 78, 87, 103, 121, 120, 101},
 			{72, 92, 95, 98, 112, 100, 103, 99}};
 	
-	private static double[][] lqt_recommended_chrominance = {
+	private static float[][] lqt_recommended_chrominance = {
 			{17, 18, 24, 47, 99, 99, 99, 99},
 			{18, 21, 26, 66, 99, 99, 99, 99},
 			{24, 26, 56, 99, 99, 99, 99, 99},
@@ -23,7 +23,7 @@ public class Quantization{
 			{99, 99, 99, 99, 99, 99, 99, 99},
 			{99, 99, 99, 99, 99, 99, 99, 99}};
 	
-	private static double[][] lqt_high_quality = {
+	private static float[][] lqt_high_quality = {
 			{1, 1, 1, 1, 1, 2, 3, 3},
 			{1, 1, 1, 1, 1, 3, 3, 3},
 			{1, 1, 1, 1, 2, 3, 3, 3},
@@ -49,7 +49,7 @@ public class Quantization{
 	
 	// Private
 	private static Matrix createLQT(int table) {
-	    double[][] lqt;
+	    float[][] lqt;
 	    if (table == 1) {
 	    	lqt = lqt_recommended_luminance;
 	    }
@@ -57,7 +57,7 @@ public class Quantization{
 	    	lqt = lqt_recommended_chrominance;
 	    }
 	    else {
-	    	lqt = new double[1][1];
+	    	lqt = new float[1][1];
 	    }
 	    Matrix result = new Matrix(lqt);
 	    return result;
